+++ packages/zowe-explorer/__tests__/__unit__/trees/dataset/DatasetActions.unit.test.ts
@@ -1934,7 +1934,7 @@ describe("Dataset Actions Unit Tests - Function hMigrateDataSet", () => {
             },
         });
 
-        await DatasetActions.hMigrateDataSet(node);
+        await DatasetActions.hMigrateDataSet(blockMocks.testDatasetTree, node);
 
         expect(migrateSpy).toHaveBeenCalledWith("HLQ.TEST.TO.NODE");
         expect(mocked(Gui.showMessage)).toHaveBeenCalled();
@@ -1953,7 +1953,7 @@ describe("Dataset Actions Unit Tests - Function hMigrateDataSet", () => {
         });
         node.contextValue = Constants.DS_DS_CONTEXT;
 
-        await DatasetActions.hMigrateDataSet(node);
+        await DatasetActions.hMigrateDataSet(blockMocks.testDatasetTree, node);
 
         expect(mocked(Gui.errorMessage)).toHaveBeenCalled();
     });
@@ -1988,7 +1988,7 @@ describe("Dataset Actions Unit Tests - Function hMigrateDataSet", () => {
             },
         });
 
-        await DatasetActions.hMigrateDataSet(node);
+        await DatasetActions.hMigrateDataSet(blockMocks.testDatasetTree, node);
 
         expect(migrateSpy).toHaveBeenCalledWith("HLQ.TEST.TO.NODE");
         expect(mocked(Gui.showMessage)).toHaveBeenCalled();
@@ -2045,7 +2045,7 @@ describe("Dataset Actions Unit Tests - Function hRecallDataSet", () => {
             },
         });
 
-        await DatasetActions.hRecallDataSet(node);
+        await DatasetActions.hRecallDataSet(blockMocks.testDatasetTree, node);
 
         expect(recallSpy).toHaveBeenCalledWith("HLQ.TEST.TO.NODE");
         expect(mocked(Gui.showMessage)).toHaveBeenCalled();
@@ -2082,7 +2082,7 @@ describe("Dataset Actions Unit Tests - Function hRecallDataSet", () => {
             },
         });
 
-        await DatasetActions.hRecallDataSet(node);
+        await DatasetActions.hRecallDataSet(blockMocks.testDatasetTree, node);
 
         expect(recallSpy).toHaveBeenCalledWith("HLQ.TEST.TO.NODE");
         expect(mocked(Gui.showMessage)).toHaveBeenCalled();